expose deploy sync api
