+++ src/commands/runtime/deploy/sync.js
@@ -11,10 +11,8 @@ governing permissions and limitations under the License.
 */
 
 const RuntimeBaseCommand = require('../../../RuntimeBaseCommand')
-const { setPaths, processPackage, deleteEntities } = require('../../../runtime-helpers')
+const { setPaths, processPackage, syncPackage } = require('../../../runtime-helpers')
 const { flags } = require('@oclif/command')
-const sha1 = require('sha1')
-const fs = require('fs')
 
 class DeploySync extends RuntimeBaseCommand {
   async run () {
@@ -32,144 +30,11 @@ class DeploySync extends RuntimeBaseCommand {
       const entities = processPackage(packages, deploymentPackages, deploymentTriggers, params)
       const ow = await this.wsk()
       const logger = this.log
-      // find project hash from server based on entities in the manifest file
-      const hashProjectSynced = await findProjectHashonServer(ow, components.projectName)
-
-      // compute the project hash from the manifest file
-      const projectHash = getProjectHash(components.manifestContent, components.manifestPath)
-      await addAnnotationsandDeploy(entities, ow, logger, components.manifestPath, components.projectName, projectHash)
-      if (projectHash !== hashProjectSynced) {
-        // delete old files with same project name that do not exist in the manifest file anymore
-        await deleteEntities(hashProjectSynced, ow, '')
-      }
+      await syncPackage(entities, ow, logger, components.projectName, components.manifestPath, components.manifestContent)
     } catch (err) {
       this.handleError('Failed to sync', err)
     }
   }
-}
-
-async function addAnnotationsandDeploy (entities, ow, logger, manifestPath, projectName, projectHash) {
-  // add whisk managed annotations
-  const opts = await ow.actions.client.options
-  const ns = opts.namespace
-  for (const pkg of entities.pkgtoCreate) {
-    const options = {}
-    options['name'] = pkg.name
-    options['package'] = {
-      annotations: [
-        {
-          key: 'whisk-managed',
-          value: {
-            file: manifestPath,
-            projectDeps: [],
-            projectHash: projectHash,
-            projectName: projectName
-          }
-        }
-      ]
-    }
-    logger(`Info: Deploying package [${pkg.name}]...`)
-    await ow.packages.update(options)
-    logger(`Info: package [${pkg.name}] has been successfully deployed.\n`)
-  }
-  for (const action of entities.actions) {
-    if (action['exec'] && action['exec']['kind']) {
-      action['exec']['components'] = action['exec']['components'].map(sequence => {
-        return `/${ns}/${sequence}`
-      })
-    }
-    action['annotations']['whisk-managed'] = {
-      file: manifestPath,
-      projectDeps: [],
-      projectHash: projectHash,
-      projectName: projectName
-    }
-    logger(`Info: Deploying action [${action.name}]...`)
-    await ow.actions.update(action)
-    logger(`Info: action [${action.name}] has been successfully deployed.\n`)
-  }
-
-  for (const trigger of entities.triggers) {
-    const managedAnnotation = {
-      key: 'whisk-managed',
-      value: {
-        file: manifestPath,
-        projectDeps: [],
-        projectHash: projectHash,
-        projectName: projectName
-      }
-    }
-    if (trigger['trigger'] && trigger['trigger']['annotations']) {
-      trigger['trigger']['annotations'].push(managedAnnotation)
-    } else {
-      trigger['trigger']['annotations'] = [managedAnnotation]
-    }
-    logger(`Info: Deploying trigger [${trigger.name}]...`)
-    await ow.triggers.update(trigger)
-    logger(`Info: trigger [${trigger.name}] has been successfully deployed.\n`)
-  }
-
-  logger('Success: Deployment completed successfully.')
-}
-
-function getProjectHash (manifestContent, manifestPath) {
-  const stats = fs.statSync(manifestPath)
-  const fileSize = stats.size.toString()
-  const hashString = `Runtime ${fileSize}\0${manifestContent}`
-  const projectHash = sha1(hashString)
-  return projectHash
-}
-
-async function findProjectHashonServer (ow, projectName) {
-  let projectHash = ''
-  const options = {}
-  // check for package with the projectName in manifest File and if found -> return the projectHash on the server
-  const resultSync = await ow.packages.list(options)
-  for (const pkg of resultSync) {
-    if (pkg.annotations.length > 0) {
-      const whiskManaged = pkg.annotations.find(elem => elem.key === 'whisk-managed')
-      if (whiskManaged !== undefined && whiskManaged.value.projectName === projectName) {
-        projectHash = whiskManaged.value.projectHash
-        return projectHash
-      }
-    }
-  }
-  // if no package exists with the projectName -> look in actions
-  const resultActionList = await ow.actions.list()
-  for (const action of resultActionList) {
-    if (action.annotations.length > 0) {
-      const whiskManaged = action.annotations.find(elem => elem.key === 'whisk-managed')
-      if (whiskManaged !== undefined && whiskManaged.value.projectName === projectName) {
-        projectHash = whiskManaged.value.projectHash
-        return projectHash
-      }
-    }
-  }
-
-  // if no action exists with the projectName -> look in triggers
-  const resultTriggerList = await ow.triggers.list()
-  for (const trigger of resultTriggerList) {
-    if (trigger.annotations.length > 0) {
-      const whiskManaged = trigger.annotations.find(elem => elem.key === 'whisk-managed')
-      if (whiskManaged !== undefined && whiskManaged.value.projectName === projectName) {
-        projectHash = whiskManaged.value.projectHash
-        return projectHash
-      }
-    }
-  }
-
-  // if no trigger exists with the projectName -> look in rules
-  const resultRules = await ow.rules.list()
-  for (const rule of resultRules) {
-    if (rule.annotations.length > 0) {
-      const whiskManaged = rule.annotations.find(elem => elem.key === 'whisk-managed')
-      if (whiskManaged !== undefined && whiskManaged.value.projectName === projectName) {
-        projectHash = whiskManaged.value.projectHash
-        return projectHash
-      }
-    }
-  }
-  return projectHash
 }
 
 DeploySync.flags = {
